app/requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
pandas>=2.0.0
numpy>=1.24.0

# NLP Processing
nltk>=3.8.0

# Explainability
lime>=0.2.0.1

# Backend - Flask
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10
gunicorn>=21.0.0

# Frontend - Streamlit
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
requests>=2.31.0

# OCR for image text extraction
easyocr>=1.7.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# Production
waitress>=2.1.2

# Testing
pytest>=7.4.0
